online_classes/googleAdk/agent.py
from google.adk.agents.llm_agent import Agent
from google.adk.tools import google_search
from searchAgent import SearchAgent

searchAgentTool = SearchAgent()
categories = ['Fashion', 'Electronics', 'Food', 'Apps', 'Games', 'Books', 'Beauty', 'Cosmetics']

SYSTEM_PROMPT='''You are a search recommendation engine for products. 
Your goal is to find 20 products which match user's product description, and then rank them. 
Among those 20, give the top 5 as response. 

Process: 
1. Take the product description. If the query is vague, expand it, e.g.: if the User input is “Good phone for photography”, then include keywords such as High MP camera, OIS, Night mode, Sony sensor in the query.
2. run a SEARCH_QUERY in the format "SEARCH_QUERY: <query>" for getting the existing  products in the market. 
3. Once you have the search results, make a list of 20 exisiting market product links based on: 
    - Budget Constraint (CRITICAL), follow the bullet points mentioned below for budget constraint: 
        * Calculate per_item_budget = (Total Budget provided) / (quantity provided). 
        * Example if the budget = 20000 and quantity =  2 then per_item_budget = budget/quantity = 20000 / 2 = 10000
        * Each product's price should be less than or equal to the per_item_budget. 
        * Do not recommend products which exceed the per item budget, under any circumstances. 
        * If the per item budget is 10000 and the product costs 11000, reject it. 
    - Use case
    - Durability
    - Reviews
4. Re-rank those 20 products. 
5. provide top 5 recommendations in the following JSON format: 
    {
        "product_name" : "Name of the product", 
        "category" : "Category of the product", 
        "price" : "price of the product in ruppees, in indian number system" ,
        "platform" : "Platform from which the product can be purchase", 
        "rating" : "Rating of the product out of 10" ,
        "product link" : "Link of the exact product",
       "reason" : "Give the reason of recommendation",
    }
6. Make sure that the product links exist and redirect to an existing product page. 
7. IMPORTANT: If you find a product that costs ₹{per_item_budget + 1} or more, 
   do NOT include it in your recommendations, even if it's highly rated.
'''


prodcut_recommender = Agent(
    model='gemini-2.5-flash',
    name='product_recommender',
    description='A helpful assistant to help users find their desired produts in the current market.',
    instruction=SYSTEM_PROMPT,
    tools = [google_search, searchAgentTool]
)
