online_classes/googleAdk/requirements.txt
google-adk==1.24.1
fastapi==0.128.7
pydantic==2.12.5
uvicorn==0.40.0
requests==2.32.5
dotenv==0.9.9